backend/requirements.txt
uvicorn
fastapi==0.71.0
fastapi-users[mongodb]==9.2.2
loguru
aiofiles
kubernetes-asyncio>=22.6.4
aiobotocore
redis>=4.2.0rc1
pyyaml
jinja2
python_on_whales
